:sparkles: :bug: :bulb: Changes in stock speed
Some significant improvements in the speed of sending text and media, also to control the status of the application.
Some recurring functions were also centralized, such as waiting for a specific activity.
I am trying to adapt the code so that it is easy to transform it into asynchronous in the future (to be able to manipulate databases and actions at the same time).